+++ document/changelog.md
@@ -1,3 +1,6 @@
+unreleased
+
+* support crypto handler version 4 (V2 and AESV2)
 
 0.0.4.0